Moves filtered map styles to const in utils
